Move particle filter into own file
